refactor: extract all business logic from EventRegistration.js into useEventRegistration hook
EventRegistration.js was 925 lines with tight coupling between:
- Data fetching (API calls, offline cache, hackathon mock data)
- Form state management (useFormValidation)
- Registration submission (queue, conflict detection, auth check)
- Rendering (success view, form view, past event view)

Extracted into src/hooks/useEventRegistration.js which returns all
state and handlers the component needs.

EventRegistration.js now only handles UI rendering:
- Loading state → spinner
- Not found → error message
- Past event → registration unavailable
- Registered → success view with confetti, calendar links, sharing
- Default → registration form with conflict modal

File reduced from 925 lines to 575 lines (-38%).
Hook is independently testable.
